Adding a new sort parameter, and method.
Adding a sorting method only include points that live in a region's own
Voronoi cell. This is mostly used for diagnostic purposes, and is not
included in the Params file. The sort method of "sort_in_vor" can be
used to perform this sorting method.

Also adding a parameter to the existing Voronoi based sorting algorithm.
This parameter allows the user to define cutoff distance of additional
points in neighboring Voronoi cells when sorting. This allows the user
to realize additional speedups by lowering the number of points a region
triangulates.