Updating Voronoi sort algorithm.
Adding a function which computes the max and average circumradius for
all triangles in a region. These are stored in each region separately,
and must be computed after the region is triangulated.

Voronoi sort now starts by including only points in the region's own
Voronoi cell. After which, it includes points that are some distance
outside the Voronoi cell by using three input parameters and the average
(or maximum) circumradius for the region.

Input parameters include:
	sort_max_radius_include -- Maximum number of circumradii to include
		in sorting
	sort_min_radius_include -- Minimum number of circumradii to include
		in sorting
	sort_max_radius_iterations -- The number of iterations to use
		sort_max_radius_include for sorting. After which
		sort_min_radius_include will be used instead.

The change from average to maximum circumradius is currently hardcoded
and only for exploratory purposes. Maximum isn't used currently because
non-delaunay triangles on the perimeter of the region could polute this
metric easier than average circumradius.

Can be used in combination with previous alpha based Voronoi sort.