Adding a heuristic "enforced" sort method.
This sort determines an approximate grid spacing for a region, based on
how many points are in the region, and what the approximate surface area
of the region is.

The approximate surface area of the region is determined by the radius
of the region/2.

Using the apprixmate grid spacing for the region, an approximate
circumdiameter is created (2*approx_spacing) and two circumdiameters are
added to the outside of the Voronoi region.

This then determines which points will be added for Triangulation.