Fixing Voronoi sort.
Updating Voronoi sort to include all points in each neighbor within some
radius. Fixes points on the side of the fine region neighbors.

Voronoi sort now includes all points which are closer to the region
center than to any of the 2nd layer neighbor centers.

This can be thought of as an expanded Voronoi cell for each region, that
extends roughly halfway through the 1st layer neighbor centers.