Adding global point update routine.
This routine is inefficient right now, and won't scale well, but might
not take a significant amount of time.

At a later point in time, it should be replaced by some custom
"all_reduce" algorithim to update the global point set.

It's used currently to update the entire point set at the end of an
iteration.